+++ src/protocol/message_process.cpp
@@ -51,6 +51,22 @@ static void send_reading_to_master(void)
 }
 #endif
 
+static void send_address_by_sms()
+{
+    ESP_LOGD(TAG, "Enviando endereço (%s) para todos os números...", settings_get_local());
+    for(uint8_t index = 0; index < settings_get_phones_list_length(); index++)
+    {
+        char *phone = settings_get_phone(index);
+
+        ESP_LOGD(TAG, "Número: %s", phone);
+        if(strcmp(phone, "") > 0)
+        {
+            sim7070g_send_sms(phone, settings_get_local());
+            vTaskDelay(pdMS_TO_TICKS(1000));
+        }
+    }
+}
+
 static void send_readings_by_sms(char *message)
 {
     ESP_LOGD(TAG, "Enviando SMS: %s", message);
@@ -61,8 +77,6 @@ static void send_readings_by_sms(char *message)
         ESP_LOGD(TAG, "Número: %s", phone);
         if(strcmp(phone, "") > 0)
         {
-            sim7070g_send_sms(phone, settings_get_local());
-            vTaskDelay(pdMS_TO_TICKS(1000));
             sim7070g_send_sms(phone, message);
             vTaskDelay(pdMS_TO_TICKS(10000));
         }
@@ -94,6 +108,7 @@ static void on_message_event_handler(protocol_data_raw_t leitura)
                 ESP_LOGI(TAG, "Função SET Litros recebida!");
                 char message[512];
                 snprintf(message, 512, "%d - %s", leitura.id, (char*)leitura.data);
+                send_address_by_sms();
                 send_readings_by_sms(message);
             }
 #endif
@@ -191,6 +206,7 @@ void get_readings_timer_callback(void *argv)
 
         char message[512];
         snprintf(message, 512, "controlador - %.2f", flowsensor_get_litros());
+        send_address_by_sms();
         send_readings_by_sms(message);
         vTaskDelay(pdMS_TO_TICKS(settings_get_interval()*60*1000));
     }